requirements.txt
python-dotenv
gradio
embedchain
torch 
google-generativeai